+++ discord/ext/vbu/cogs/help_command.py
@@ -16,7 +16,7 @@ class Help(vbu.Cog):
     def cog_unload(self):
         self.bot.help_command = self._original_help_command
 
-    @vbu.command(name="commands", hidden=True, add_slash_command=False)
+    @commands.command(name="commands", hidden=True, add_slash_command=False)
     async def _commands(self, ctx: vbu.Context, *args):
         """
         An alias for help.
@@ -24,7 +24,7 @@ class Help(vbu.Cog):
 
         return await ctx.send_help(*args)
 
-    @vbu.command(hidden=True, add_slash_command=False)
+    @commands.command(hidden=True, add_slash_command=False)
     @commands.has_permissions(manage_messages=True)
     async def channelhelp(self, ctx: vbu.Context, *args):
         """