+++ discord/ext/vbu/cogs/utils/converters/__init__.py
@@ -5,3 +5,15 @@ from .boolean_converter import BooleanConverter
 from .colour_converter import ColourConverter
 from .filtered_user import FilteredUser, FilteredMember
 from .number_converter import NumberConverter
+
+
+__all__ = (
+    'UserID',
+    'ChannelID',
+    'EnumConverter',
+    'BooleanConverter',
+    'ColourConverter',
+    'FilteredUser',
+    'FilteredMember',
+    'NumberConverter',
+)