+++ discord/ext/vbu/cogs/utils/errors.py
@@ -3,8 +3,22 @@ from .checks.meta_command import InvokedMetaCommand
 from .checks.bot_is_ready import BotNotReady
 from .checks.is_voter import IsNotVoter
 from .checks.is_bot_support import NotBotSupport
-# from .checks.slash_commands import IsSlashCommand, IsNotSlashCommand, BotNotInGuild
 from .checks.is_upgrade_chat_subscriber import IsNotUpgradeChatSubscriber, IsNotUpgradeChatPurchaser
 from .missing_required_argument import MissingRequiredArgumentString
 from .time_value import InvalidTimeDuration
 from .menus.errors import ConverterFailure, ConverterTimeout
+
+
+__all__ = (
+    'ConfigNotSet',
+    'InvokedMetaCommand',
+    'BotNotReady',
+    'IsNotVoter',
+    'NotBotSupport',
+    'IsNotUpgradeChatSubscriber',
+    'IsNotUpgradeChatPurchaser',
+    'MissingRequiredArgumentString',
+    'InvalidTimeDuration',
+    'ConverterFailure',
+    'ConverterTimeout',
+)